+++ common/src/main/java/com/dudko/blazinghot/content/casting/Molds.java
@@ -4,7 +4,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 import com.dudko.blazinghot.BlazingHot;
-import com.dudko.blazinghot.content.item.MoldItem;
 import com.dudko.blazinghot.multiloader.BlazingBuilderTransformers;
 import com.dudko.blazinghot.multiloader.fluid.MultiAmount;
 import com.dudko.blazinghot.registry.BlazingTags;
@@ -14,7 +13,6 @@ import com.tterrag.registrate.util.entry.ItemEntry;
 
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
 
 public class Molds {
 
@@ -24,26 +22,16 @@ public class Molds {
 
 	public static final Mold BLANK = new Mold("blank").register();
 	public static final Mold INGOT = new Mold("ingot", MultiAmount.INGOT.get(), Items.INGOTS.tag()).register();
-	public static final Mold
-			NUGGET =
-			new Mold("nugget", MultiAmount.NUGGET.get(), Items.NUGGETS.tag()).register();
+	public static final Mold NUGGET = new Mold("nugget", MultiAmount.NUGGET.get(), Items.NUGGETS.tag()).register();
 	public static final Mold SHEET = new Mold("sheet", MultiAmount.INGOT.get(), Items.PLATES.tag()).register();
 	public static final Mold ROD = new Mold("rod", MultiAmount.ROD.get(), Items.RODS.tag()).register();
-
-	public static long getMoldCapacity(ItemStack stack) {
-		if (stack == null || stack.isEmpty()) return 2 * MultiAmount.INGOT.get();
-		else if (stack.getItem() instanceof MoldItem moldItem) {
-			return moldItem.capacity;
-		}
-		else return MultiAmount.INGOT.get();
-	}
 
 	public static class Mold {
 
 		public final String name;
 		public final long capacity;
 		public final TagKey<Item> shape;
-		public final Map<MoldType, ItemEntry<MoldItem>> items = new HashMap<>();
+		public final Map<MoldType, ItemEntry<? extends Item>> items = new HashMap<>();
 
 		public Mold(String name) {
 			this(name, 0, null);
@@ -63,8 +51,7 @@ public class Molds {
 			for (MoldType type : MoldType.values()) {
 				items.put(type,
 						registrate
-								.item(type.name + "_" + name + "_mold",
-										p -> new MoldItem(p, capacity, shape, type.reusable))
+								.item(type.name + "_" + name + "_mold", Item::new)
 								.transform(BlazingBuilderTransformers.mold(name, type))
 								.register());
 			}
@@ -72,7 +59,7 @@ public class Molds {
 			return this;
 		}
 
-		public ItemEntry<MoldItem> get(MoldType type) {
+		public ItemEntry<? extends Item> get(MoldType type) {
 			return items.get(type);
 		}